chsmack: Add filtering options
This commit add filtering options to the command chsmack.
The filtering options are allowing to select files to list
or to modify according to values of their attributes.

Example: the command "chsmack --if-exec spurious -E -a Nobody -r /"
set the Smack access property to Nobody and drops the exec property
of any file that have the exec label spurious.

Added to that filtering options, the new option --name-only allows
to get the list of files matching a criteria.

See manual of chsmack for details.

Signed-off-by: José Bollo <jose.bollo@iot.bzh>